feat(ai): complete DetThenRec feature(not tested)
additional:
- define bounding box struct
