+++ .gitignore
@@ -1,2 +1,3 @@
 dataset/
 .env
+packages/
