+++ ai/Cargo.toml
@@ -24,6 +24,7 @@ face-detection = { path = "./face-detection/" }
 image = "0.25"
 imageproc = "0.25.0"
 mysql = "25.0.1"
+serde = { version = "1.0.214", features = ["derive"] }
 tokio = { version = "1.40.0", features = ["rt-multi-thread", "signal", "tracing", "process"] }
 tracing = "0.1.40"
 tracing-subscriber = "0.3.18"
