+++ ai/src/model/utils.rs
@@ -1,16 +1,15 @@
 use ab_glyph::{FontRef, PxScale};
 use image::{Rgb, RgbImage};
-use imageproc::{
-    drawing::{draw_hollow_rect, draw_text},
-    rect::Rect,
-};
+use imageproc::drawing::{draw_hollow_rect_mut, draw_text_mut};
+
+use super::bounding_box::BoundingBox;
 
 type DlibRectangle = dlib_face_recognition::Rectangle;
 
 pub trait LabelID {
     fn label(
         &mut self,
-        rect: &DlibRectangle,
+        rect: &BoundingBox,
         name: &str,
         font: &FontRef,
         font_height: f32,
@@ -21,14 +20,14 @@ pub trait LabelID {
 impl LabelID for RgbImage {
     fn label(
         &mut self,
-        rect: &DlibRectangle,
+        rect: &BoundingBox,
         name: &str,
         font: &FontRef,
         font_height: f32,
         scale: PxScale,
     ) {
-        draw_hollow_rect(self, Rectangle(rect).into(), Rgb([255, 0, 0]));
-        draw_text(
+        draw_hollow_rect_mut(self, rect.into(), Rgb([255, 0, 0]));
+        draw_text_mut(
             self,
             Rgb([255, 0, 0]),
             rect.left,
@@ -37,16 +36,5 @@ impl LabelID for RgbImage {
             font,
             name,
         );
-    }
-}
-
-struct Rectangle<'a>(&'a DlibRectangle);
-
-impl<'a> Into<Rect> for Rectangle<'a> {
-    fn into(self) -> Rect {
-        Rect::at(self.0.left, self.0.top).of_size(
-            (self.0.right - self.0.left) as u32,
-            (self.0.bottom - self.0.top) as u32,
-        )
     }
 }
